bit more refactor
